Rename project.modelcomponents.fields to project.fields.
"Field" classes have never belonged as part of the 'project.models'
module, and this causes problems with import dependencies, especially
as SafeHTMLField is now used extensively outside the 'projects' app.

Rename this module to clearly separate the field class - which is
standalone and reusable - from the model classes that are defined in
project/modelcomponents/.
