+++ physionet-django/project/modelcomponents/legacy.py
@@ -5,9 +5,9 @@ from django.db import models
 from django.utils import timezone
 import pytz
 
+from project.fields import SafeHTMLField
 from project.modelcomponents.access import License
 from project.modelcomponents.coreproject import CoreProject, ProjectType
-from project.modelcomponents.fields import SafeHTMLField
 from project.modelcomponents.metadata import PublishedPublication, Contact
 from project.modelcomponents.publishedproject import PublishedProject
