+++ physionet-django/project/modelcomponents/metadata.py
@@ -9,8 +9,8 @@ from django.db import models
 from django.utils import timezone
 from django.utils.html import escape, format_html, mark_safe
 from html2text import html2text
+from project.fields import SafeHTMLField
 from project.modelcomponents.access import AccessPolicy, AnonymousAccess
-from project.modelcomponents.fields import SafeHTMLField
 from project.modelcomponents.authors import Affiliation
 from project.utility import LinkFilter, get_directory_info, get_file_info, list_items
 from project.validators import validate_title, validate_topic, validate_version
