+++ physionet-django/project/modelcomponents/publishedproject.py
@@ -11,9 +11,9 @@ from django.utils import timezone
 from django.utils.text import slugify
 
 from notification.models import News
+from project.fields import SafeHTMLField
 from project.managers.publishedproject import PublishedProjectManager
 from project.modelcomponents.access import DataAccessRequest, DataAccessRequestReviewer, DUASignature
-from project.modelcomponents.fields import SafeHTMLField
 from project.modelcomponents.metadata import Metadata, PublishedTopic
 from project.modelcomponents.submission import SubmissionInfo
 from project.modelcomponents.access import AccessPolicy
