+++ physionet-django/project/models.py
@@ -2,7 +2,6 @@ from project.modelcomponents.access import *
 from project.modelcomponents.activeproject import *
 from project.modelcomponents.authors import *
 from project.modelcomponents.coreproject import *
-from project.modelcomponents.fields import *
 from project.modelcomponents.legacy import *
 from project.modelcomponents.metadata import *
 from project.modelcomponents.projecttype import *
